requirements.txt
PyQt5>=5.15.0
Pillow>=10.0.0
PyMuPDF>=1.23.0

# Dev / build
pytest>=7.0.0
pyinstaller>=6.0.0
